# documentation/api-documentation.yml
openapi: "3.0.0"
info:
  version: "3.0.1"
  title: "FemDevs API Documentation"
  description: |-
    The documentation for the FemDevs API.
    Do note that some routes have permission-based requirements that your user must be quilified for in order to access them.
    Check the description for each route in case you are unsure.

    Permissions Overrides:

    The following permissions will bypass all permissions unless its marked otherwise:
    * `Global::Role.Developer`
    * `Global::Role.Administrator`
    * `Global::Role.Owner`
  termsOfService: "https://legal.thefemdevs.com/terms"
  contact:
    name: "FemDevs"
    url: "https://thefemdevs.com"
    email: "support@thefemdevs.com"
  license:
    name: "GNU Affero General Public License v3.0"
    url: "https://www.gnu.org/licenses/agpl-3.0.en.html"
servers:
  - url: "https://api.thefemdevs.com/"
    description: "Production Server"
paths:
...
  /crypto/enc:
    get:
      operationId: "CryptographyEnc"
      description: |-
        API Endpoint for encrypting data

        Permissions Required: `Cryptography::Encrypt`
      summary: "Encryption Endpoint"
      parameters:
        - in: "query"
          name: "data"
          description: "The data to encrypt"
          example:
            value: "testing"
          schema:
            type: string
      responses:
        200:
          description: "The Encrypted Data"
          content:
            application/json:
              schema:
                type: string
              examples:
                data:
                  summary: "Encrypted Data"
                  value: "Pc6W8_h9XSg8LVhZ8yAP4vtui0sJIMNmC_E18ee9JBQyEmBPLZBDz5Jx745u31Pbr2EkS9Z2S8oFB02uwzpWtt3PH48HE-Wj27VkpvV1glrpZ0f_cW0Ema-gT-6S675xCgb9cy0ZTx3rH-kzuVZXaOwrMjMJTVe7ZpyH8zyUNHnDo0xvkhq9Z523azZm1mq1rcJWMJl8iCuZc6TQp5E2-4xiAZF59ZCZyQCtzxryPzX6aR5ENBvBVdkOaNWtmjWc9ZM5Ni9PXc2cVGbi5IDo6RvX5KuU2DUvGpIKvf15xLptPXZPLNtU4syXjXI4vA03EpV1-Dqs6H7FrtXLAu9M5Q7NxaSxynr4oyJC7CzIXJtJVUwrFN9I-zSJCcR9rE90XzolaWICUjfhSkMtPBTkr_ibNFcyPhG-oaMPGuc1Hb8gJzEDDtWTGPZdV2rbGJIw3XBi_h80fj8F_DQujSGsbo2DmBwSF-PZtKG5WHK9cf3rWCIxF1Sc5Hhkd-HnhrFrSv30Kjvede88MmofhchsS9cEn4eFF7rfj35blLBxGKBKzhS0k_k7FWQznK6J24lfUR_Gtkz3IJlTcqQN-ixucjHdI58xheVETAg_PkK3KZIwgabxQRSz96JUANyEPIPdodIF2f4PpEMomjYp8mCc_0F26uUSgsm9_LcJvC9oNZU"
        400:
          description: "Error With Encryption (Client Caused)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                no-data:
                  summary: "Missing Data Parameter"
                  value:
                    code: 6
                    message: "Missing Query Parameter(s)"
                invalid-data:
                  summary: "Invalid Data Parameter"
                  value:
                    code: 7
                    message: "Invalid Query Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        500:
          description: "Error With Encryption (Server Caused)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 12
                    message: "Unable to encrypt data"
      security:
        - APIToken: []
  /crypto/dec:
    get:
      operationId: "CryptographyDecr"
      description: |-
        API Endpoint for decrypting data

        Permissions Required: `Cryptography::Decrypt`
      summary: "Decryption Endpoint"
      parameters:
        - in: "query"
          name: "data"
          description: "The data to decrypt"
          schema:
            type: string
          example:
            value: 'Pc6W8_h9XSg8LVhZ8yAP4vtui0sJIMNmC_E18ee9JBQyEmBPLZBDz5Jx745u31Pbr2EkS9Z2S8oFB02uwzpWtt3PH48HE-Wj27VkpvV1glrpZ0f_cW0Ema-gT-6S675xCgb9cy0ZTx3rH-kzuVZXaOwrMjMJTVe7ZpyH8zyUNHnDo0xvkhq9Z523azZm1mq1rcJWMJl8iCuZc6TQp5E2-4xiAZF59ZCZyQCtzxryPzX6aR5ENBvBVdkOaNWtmjWc9ZM5Ni9PXc2cVGbi5IDo6RvX5KuU2DUvGpIKvf15xLptPXZPLNtU4syXjXI4vA03EpV1-Dqs6H7FrtXLAu9M5Q7NxaSxynr4oyJC7CzIXJtJVUwrFN9I-zSJCcR9rE90XzolaWICUjfhSkMtPBTkr_ibNFcyPhG-oaMPGuc1Hb8gJzEDDtWTGPZdV2rbGJIw3XBi_h80fj8F_DQujSGsbo2DmBwSF-PZtKG5WHK9cf3rWCIxF1Sc5Hhkd-HnhrFrSv30Kjvede88MmofhchsS9cEn4eFF7rfj35blLBxGKBKzhS0k_k7FWQznK6J24lfUR_Gtkz3IJlTcqQN-ixucjHdI58xheVETAg_PkK3KZIwgabxQRSz96JUANyEPIPdodIF2f4PpEMomjYp8mCc_0F26uUSgsm9_LcJvC9oNZU'
      responses:
        200:
          description: "The Decrypted Data"
          content:
            application/json:
              schema:
                type: string
              examples:
                data:
                  summary: "Decrypted Data"
                  value: "testing"
        400:
          description: "Error with Decryption (Client Caused)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                no-data:
                  summary: "Missing Data Parameter"
                  value:
                    code: 6
                    message: "Missing Query Parameter(s)"
                invalid-data:
                  summary: "Invalid Data Parameter"
                  value:
                    code: 7
                    message: "Invalid Query Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        500:
          description: "Error with Decryption (Server Caused)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 13
                    message: "Unable to decrypt data"
      security:
        - APIToken: []
...
  /dictionary/def:
    get:
      operationId: "GetDictionaryDefinition"
      description: |-
        Gets the definition for the given word.
        Returns the definition if the word exists, otherwise returns an error.

        Permissions Required: `Dictionary::Definition`
      summary: "Get Dictionary Definition"
      parameters:
        - in: "header"
          name: "x-word"
          description: "The word to get the definition for"
          schema:
            type: string
      responses:
        200:
          description: "The definition for the given word"
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  type: string
              example:
                cool:
                  value:
                    - "moderately cold : lacking in warmth"
                    - "marked by steady dispassionate calmness and self-control"
                    - "lacking ardor or friendliness"
        400:
          description: "Missing word information"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Word:
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Invalid-Word:
                  value:
                    code: 9
                    message: "Invalid Header Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given word"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retreive requested information"
      security:
        - APIToken: []
  /dictionary/syn:
    get:
      operationId: "GetDictionarySynonyms"
      description: |-
        Gets the synonyms for the given word.
        Returns the synonyms if the word exists, otherwise returns an error.

        Permissions Required: `Dictionary::Synonym`
      summary: "Get Dictionary Synonyms"
      parameters:
        - in: "header"
          name: "x-word"
          description: "The word to get the synonyms for"
          schema:
            type: string
      responses:
        200:
          description: "The synonyms for the given word"
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  type: string
              examples:
                smile:
                  value:
                    syns:
                      - beam
                      - grin
                      - laugh
                      - sneer
                      - snicker
                      - snigger
        400:
          description: "Missing word information"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Word:
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Invalid-Word:
                  value:
                    code: 9
                    message: "Invalid Header Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given word"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retreive requested information"
      security:
        - APIToken: []
  /dictionary/ant:
    get:
      operationId: "GetDictionaryAntonyms"
      description: |-
        Gets the antonyms for the given word.
        Returns the antonyms if the word exists, otherwise returns an error.

        Permissions Required: `Dictionary::Antonym`
      summary: "Get Dictionary Antonyms"
      parameters:
        - in: "header"
          name: "x-word"
          description: "The word to get the antonyms for"
          schema:
            type: string
      responses:
        200:
          description: "The antonyms for the given word"
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  type: string
              examples:
                happy:
                  value:
                    ants:
                      - "hapless"
                      - "ill-fated"
                      - "ill-starred"
                      - "luckless"
                      - "star-crossed"
                      - "unfortunate"
                      - "unhappy"
                      - "unlucky"
                      - "displeased"
                      - "dissatisfied"
                      - "joyless"
                      - "sad"
                      - "unhappy"
                      - "unpleased"
                      - "unsatisfied"
                      - "discontent"
                      - "discontented"
                      - "displeased"
                      - "dissatisfied"
                      - "malcontent"
                      - "malcontented"
                      - "unhappy"
                      - "hapless"
                      - "ill-fated"
                      - "ill-starred"
                      - "luckless"
                      - "snakebit"
                      - "star-crossed"
                      - "unfortunate"
                      - "unhappy"
                      - "unlucky"
                      - "improper"
                      - "inapplicable"
                      - "inapposite"
                      - "inappropriate"
                      - "inapt"
                      - "incongruous"
                      - "indecent"
                      - "infelicitous"
                      - "malapropos"
                      - "misbecoming"
                      - "unapt"
                      - "unbecoming"
                      - "unbeseeming"
                      - "unfit"
                      - "unfitting"
                      - "unhappy"
                      - "unmeet"
                      - "unseemly"
                      - "unsuitable"
                      - "wrong"
        400:
          description: "Missing word information"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Word:
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Invalid-Word:
                  value:
                    code: 9
                    message: "Invalid Header Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given word"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retreive requested information"
      security:
        - APIToken: []
...
components:
  securitySchemes:
    APIToken:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: |-
        JSON Web Token Authentication with information containing your license key

        Permissions:
        - `Global::Data.Read` (0:0:0) - **Unused**
        - `Global::Data.Write` (0:0:1) - **Unused**
        - `Global::Data.Create` (0:0:2) - **Unused**
        - `Global::Data.Delete` (0:0:3) - **Unused**
        - `Global::Token.Read` (0:1:0) - Allows reading of token information
        - `Global::Token.ReadAll` (0:1:1) - Allows reading of all token information
        - `Global::Token.Write` (0:1:2) - **Unused**
        - `Global::Token.Create` (0:1:3) - Allows creating of token information
        - `Global::Token.Delete` (0:1:4) - Allows deleting of token 
        - `Global::User.Read` (0:2:0) - Read user information
        - `Global::User.ReadAll` (0:2:1) - **Unused**
        - `Global::User.Write` (0:2:2) - Write/Update user information
        - `Global::User.Create` (0:2:3) - Create a user profile
        - `Global::User.Delete` (0:2:4) - Delete a user profile
        - `Cryptography::Encrypt` (1:0) - Ability to encrypt data using the `/crypto/enc` endpoint
        - `Cryptography::Decrypt` (1:1) - Ability to decrypt data using the `/crypto/dec` endpoint
        - `Dictionary::Definition` (2:0) - Ability to get the definition of a word using the `/dict/def` endpoint
        - `Dictionary::Synonym` (2:1) - Ability to get the synonyms of a word using the `/dict/syn` endpoint
        - `Dictionary::Antonym` (2:2) - Ability to get the antonyms of a word using the `/dict/ant` endpoint
        - `Facts::Dog` (3:0) - Ability to get a random dog fact using the `/facts/dog` endpoint
        - `Facts::Cat` (3:1) - Ability to get a random cat fact using the `/facts/cat` endpoint
        - `Facts::Random` (3:2) - Ability to get a random fact using the `/facts/random` endpoint
        - `Location::Coord` (4:0) - Ability to get the location of coordinates using the `/location/coords` endpoint
        - `Location::Pluscode` (4:1) - Ability to get the location of a pluscode using the `/location/pluscode` endpoint
        - `Location::Address` (4:2) - Ability to get the location of an address using the `/location/address` endpoint
        - `Phone::Lookup` (5:0) - Ability to get the details of a phone number using the `/phone/details` endpoint
        - `Phone::Validate` (5:1) - **Unused**
        - `Quote::Kanye` (6:0) - Ability to get a random Kanye West quote using the `/quotes/kanye` endpoint
        - `Quote::RonSwanson` (6:1) - Ability to get a random Ron Swanson quote using the `/quotes/ronswanson` endpoint
        - `Quote::Random` (6:2) - Ability to get a random quote using the `/quotes/random` endpoint
        - `Weather::Current` (7:0) - Ability to get the current weather using the `/weather/current` endpoint
        - `Minecraft::Hive.Player` (8:0:0) - Ability to get information about a player on the Hive Minecraft server using the `/minecraft/hive/player` endpoint
...
  schemas:
...
    ErrorData:
      description: |-
        The error data that is returned when an error occurs

        All Error Codes are located in `/functions/errorResponse.js` [GitHub](https://github.com/femdevs/femdev-website/blob/main/functions/errorResponse.js)
      required:
        - code
        - message
      properties:
        code:
          nullable: false
          type: integer
          description: "The error code"
        message:
          nullable: false
          type: string
          description: "The error message"
...